Tutorial2: Adding an image to the window
This update adds the asset directory to the project and uses two
surfaces and a window. The gWindow is the main window interface. The
window contains a surface which is obtained and stored in
gScreenSurface. Another surface is created called gHelloWorld, a bmp
file (BitMap1.bmp) is loaded onto gHelloWorld. In the main function the
gHelloWorld surface is placed ontop of the gScreenSurface.

Along with this, init(), load(), and close() functions were created.
This creates a bit of modularity in the program and allows us to have a
cleaner main method and more organized code.